Refactor interface DTO's and rut id to long

+++ TallerJava/src/main/java/org/tallerJava/purchaseModule/domain/repo/CommerceRepository.java
@@ -3,5 +3,5 @@ package org.tallerJava.purchaseModule.domain.repo;
 import org.tallerJava.purchaseModule.domain.PurchaseCommerce;
 
 public interface CommerceRepository {
-    public PurchaseCommerce findByRut(int rut);
+    public PurchaseCommerce findByRut(long rut);
 }
